# Fix canonifyurl vs schemaless links