@bep 2015-02-17
@spf13 just fixed a potential issue (haven't experienced it while testing, but just to be sure) with this PR, amended.